+++ example/lib/app/app_module.dart
@@ -4,6 +4,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'app_bloc.dart';
 import 'app_widget.dart';
 import 'modules/home/home_module.dart';
+import 'modules/tabs/tabs_module.dart';
 
 class AppModule extends MainModule {
   @override
@@ -13,7 +14,8 @@ class AppModule extends MainModule {
 
   @override
   List<Router> get routers => [
-        Router(Modular.initialRoute, module: HomeModule()),
+        Router(Modular.initialRoute, module: TabsModule()),
+        // Router(Modular.initialRoute, module: HomeModule()),
       ];
 
   @override