+++ example/lib/app/modules/tabs/modules/tab1/pages/page2/page2_page.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class Page2Page extends StatefulWidget {
+  final String title;
+  const Page2Page({Key key, this.title = "Page2"}) : super(key: key);
+
+  @override
+  _Page2PageState createState() => _Page2PageState();
+}
+
+class _Page2PageState extends State<Page2Page> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: RaisedButton(onPressed: () {
+          Navigator.of(context).pushNamed('/page1');
+        }),
+      ),
+    );
+  }
+}