+++ example/lib/app/modules/tabs/modules/tab1/tab1_bloc.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_modular/flutter_modular.dart';
+
+class Tab1Bloc extends Disposable {
+  //dispose will be called automatically by closing its streams
+  @override
+  void dispose() {}
+}