+++ example/test/app/modules/tabs/modules/tab1/pages/page1/page1_page_test.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_modular/flutter_modular_test.dart';
+
+import 'package:example/app/modules/tabs/modules/tab1/pages/page1/page1_page.dart';
+
+main() {
+  testWidgets('Page1Page has title', (WidgetTester tester) async {
+    //  await tester.pumpWidget(buildTestableWidget(Page1Page(title: 'Page1')));
+    //  final titleFinder = find.text('Page1');
+    //  expect(titleFinder, findsOneWidget);
+  });
+}