+++ example/test/app/modules/tabs/modules/tab1/pages/page2/page2_page_test.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_modular/flutter_modular_test.dart';
+
+import 'package:example/app/modules/tabs/modules/tab1/pages/page2/page2_page.dart';
+
+main() {
+  testWidgets('Page2Page has title', (WidgetTester tester) async {
+    //  await tester.pumpWidget(buildTestableWidget(Page2Page(title: 'Page2')));
+    //  final titleFinder = find.text('Page2');
+    //  expect(titleFinder, findsOneWidget);
+  });
+}