+++ example/test/app/modules/tabs/modules/tab1/tab1_page_test.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:flutter_modular/flutter_modular_test.dart';
+
+import 'package:example/app/modules/tabs/modules/tab1/tab1_page.dart';
+
+main() {
+  testWidgets('Tab1Page has title', (WidgetTester tester) async {
+    //  await tester.pumpWidget(buildTestableWidget(Tab1Page(title: 'Tab1')));
+    //  final titleFinder = find.text('Tab1');
+    //  expect(titleFinder, findsOneWidget);
+  });
+}