+++ lib/flutter_modular.dart
@@ -13,4 +13,5 @@ export 'src/widgets/module_widget.dart';
 export 'src/widgets/consumer_widget.dart';
 export 'src/widgets/modular_stateless_widget.dart';
 export 'src/widgets/modular_stateful_widget_state.dart';
+export 'src/widgets/router_outlet.dart';
 export 'src/exceptions/modular_error.dart';